clean and refacto filters
