+++ geotrek/cirkwi/filters.py
@@ -1,25 +1,8 @@
-from django_filters import ModelMultipleChoiceFilter, FilterSet
-from django_filters.fields import ModelChoiceField
+from django_filters import FilterSet
 from geotrek.authent.models import Structure
+from geotrek.common.filters import ComaSeparatedMultipleModelChoiceFilter
 from geotrek.common.models import TargetPortal
 from geotrek.trekking.models import POI, Trek
-from django.forms import ValidationError
-
-
-class ComaSeparatedMultipleModelChoiceField(ModelChoiceField):
-    def to_python(self, value):
-        if value in self.empty_values:
-            return None
-        try:
-            key = self.to_field_name or 'pk'
-            value = self.queryset.filter(**{f'{key}__in': value.split(',')})
-        except (ValueError, TypeError):
-            raise ValidationError(self.error_messages['invalid_choice'], code='invalid_choice')
-        return value
-
-
-class ComaSeparatedMultipleModelChoiceFilter(ModelMultipleChoiceFilter):
-    field_class = ComaSeparatedMultipleModelChoiceField
 
 
 class CirkwiPOIFilterSet(FilterSet):
