+++ geotrek/tourism/filters.py
@@ -27,12 +27,8 @@ class TouristicContentFilterSet(ZoningFilterSet, StructureRelatedFilterSet):
         field_name='provider',
         empty_label=_("Provider"),
         label=_("Provider"),
-        choices=(('', '---------'),)
+        choices=lambda: TouristicContent.objects.provider_choices()
     )
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.form.fields['provider'].choices = TouristicContent.objects.provider_choices()
 
     class Meta(StructureRelatedFilterSet.Meta):
         model = TouristicContent
@@ -73,12 +69,8 @@ class TouristicEventFilterSet(ZoningFilterSet, StructureRelatedFilterSet):
         field_name='provider',
         empty_label=_("Provider"),
         label=_("Provider"),
-        choices=(('', '---------'),)
+        choices=lambda: TouristicEvent.objects.provider_choices()
     )
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.form.fields['provider'].choices = TouristicEvent.objects.provider_choices()
 
     class Meta(StructureRelatedFilterSet.Meta):
         model = TouristicEvent
